feat: [#188887754 | AB#8670] Remove duplicate close business task and the anytime action link category
Previously, there were two anytime actions for closing a business. The "Close Your Business" option was task, and the "View the Steps to Close Your Business" option was a link directly to https://business.nj.gov/pages/closing-your-business.

This commit removes the "View the Steps to Close Your Business" option. It also removes the entire link anytime actions collection and associated logic, because the deleted option was the only link and we do not anticipate more being added.