feat: [AB#12716] Enable reproducing test randomness in Cypress
This commit
- For each Cypress tests, chooses a random seed, and seeds all subsequent calls to `Math.random()` in the test with the seed
- Logs this randomly chosen seed to the console if the test fails
- Documents in the README how to find the random seed and use it to run tests locally

It also skips `multiple-businesses.spec.ts` as we've discussed about skipping flaky tests.

Approach
- Seeds each test individually with a different random seed in a `beforeEach`, and logs if a test fails in an `AfterEach`

This approach differs from those taken for unit tests because the cypress test runner mocha, [unlike the unit test runner jest](https://stackoverflow.com/questions/42000137/check-if-test-failed-in-aftereach-of-jest), is able to determine in a `AfterEach` whether the test has failed. Additionally, unlike for unit tests, this does not enforce a name-uniqueness constraint on the map because cypress tests are re-run on the CI when they fail.

This commit resolves #[12716](https://dev.azure.com/NJInnovation/Business%20First%20Stop/_workitems/edit/12716).