chore: [AB#7782] Encrypt/Decrypt Prefactor for Tax Pin
This commit renames names around `AWSEncryptionDecryptionFactory`, the `/api/decrypt` endpoint, and the apiClient call to the endpoint to more generically refer to encryption and decryption without specifying `tax`. Does not touch anything higher-level than that, e.g. `getDecryptedTaxId` on the front or `getTaxId` on the back.

This commit contributes to #[7782](https://dev.azure.com/NJInnovation/Business%20First%20Stop/_workitems/edit/7782).