feat: [AB#7782] Encrypt profile tax pin migration
This commit adds a migration that
1. Adds an `encryptedTaxPin` field that is an encryption of the profile `taxPin` field if defined
2. Masks the the profile `taxPin` field with `*`

We are okay doing this encryption on a rolling per-user basis, because likely not many users have put in their tax pin, en we have ongoing work to run migrations for all users.

This is the first migration that makes an async call within the migration function. Wrapping a `Promise.resolve()` turns all migration function return values into a promise that is then awaited.

This also makes the assumption that no profile `taxPin`s are masked as of v160.

This commit contributes to #[7782](https://dev.azure.com/NJInnovation/Business%20First%20Stop/_workitems/edit/7782).