feat: [AB#7782] Encrypt Profile Tax Pin (#10273)
* feat: [AB#7782] Encrypt profile tax pin migration

This commit adds a migration that
1. Adds an `encryptedTaxPin` field that is an encryption of the profile `taxPin` field if defined
2. Masks the the profile `taxPin` field with `*`

We are okay doing this encryption on a rolling per-user basis, because likely not many users have put in their tax pin, en we have ongoing work to run migrations for all users.

This is the first migration that makes an async call within the migration function. Wrapping a `Promise.resolve()` turns all migration function return values into a promise that is then awaited.

This also makes the assumption that no profile `taxPin`s are masked as of v160.

This commit contributes to #[7782](https://dev.azure.com/NJInnovation/Business%20First%20Stop/_workitems/edit/7782).

* feat: [AB#7782] Encrypt profile tax pin on user save and decrypt in frontend component

This commit
- Adds masking and encrypting the profile tax pin, if not yet masked and encrypted, to userRouter (this is the only backend change)
- Adds functionality to the `TaxPin` component to decrypt tax pin if encrypted and toggle show/hide
- Adds `encryptedTaxId` and `encryptedTaxPin` fields on `taxClearanceCertificateData`, so they can be populated by default by profile data to be provided to the component for decryption
    - Encrypting the taxId and taxPin fields on tax clearance certificate will be handled in a separate ticket (https://dev.azure.com/NJInnovation/Business%20First%20Stop/_workitems/edit/12793)

It also
- Fixes a tiny bug in `TaxId` component show/hide, where if a user saves their tax ID, shows and deletes it, clicks to e.g. "Permits & Licenses" then clicks back, clicking "Show" does nothing and they cannot change the field without a refresh.

This commit resolves #[7782](https://dev.azure.com/NJInnovation/Business%20First%20Stop/_workitems/edit/7782).